third_party/WebKit/LayoutTests/inspector/split-view-expected.txt

Tests how split view saving to settings works.

Create default split view
Running split view test with the following parameters: useFraction = false, shouldSaveShowMode = false
Creating split view
    Sidebar size = 101, showMode = Both, vertical
    Setting value: undefined
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":101}}
Showing sidebar
    Sidebar size = 101, showMode = Both, vertical
    Setting value: {"vertical":{"size":101}}
Resizing
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":201}}
Showing sidebar
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201}}
Making horizontal
    Sidebar size = 102, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201}}
Resizing
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Recreating split view
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Making horizontal
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Showing sidebar
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Recreating split view
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}
Making horizontal
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201},"horizontal":{"size":202}}

Create default split view
Running split view test with the following parameters: useFraction = false, shouldSaveShowMode = true
Creating split view
    Sidebar size = 101, showMode = Both, vertical
    Setting value: {"vertical":{"size":101,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":101,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 101, showMode = Both, vertical
    Setting value: {"vertical":{"size":101,"showMode":"Both"}}
Resizing
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201,"showMode":"Both"}}
Making horizontal
    Sidebar size = 102, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"Both"},"horizontal":{"size":102,"showMode":"Both"}}
Resizing
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"Both"},"horizontal":{"size":202,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"Both"},"horizontal":{"size":202,"showMode":"OnlyMain"}}
Recreating split view
    Sidebar size = 201, showMode = Both, vertical
    Setting value: {"vertical":{"size":201,"showMode":"Both"},"horizontal":{"size":202,"showMode":"OnlyMain"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"},"horizontal":{"size":202,"showMode":"OnlyMain"}}
Making horizontal
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"},"horizontal":{"size":202,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"},"horizontal":{"size":202,"showMode":"Both"}}
Recreating split view
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"},"horizontal":{"size":202,"showMode":"Both"}}
Making horizontal
    Sidebar size = 202, showMode = Both, horizontal
    Setting value: {"vertical":{"size":201,"showMode":"OnlyMain"},"horizontal":{"size":202,"showMode":"Both"}}

Create default split view
Running split view test with the following parameters: useFraction = true, shouldSaveShowMode = false
Creating split view
    Sidebar size = 115, showMode = Both, vertical
    Setting value: undefined
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.23}}
Showing sidebar
    Sidebar size = 115, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.23}}
Resizing
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.25}}
Showing sidebar
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25}}
Making horizontal
    Sidebar size = 120, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25}}
Resizing
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Recreating split view
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Making horizontal
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Showing sidebar
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Recreating split view
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}
Making horizontal
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25},"horizontal":{"size":0.26}}

Create default split view
Running split view test with the following parameters: useFraction = true, shouldSaveShowMode = true
Creating split view
    Sidebar size = 115, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.23,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.23,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 115, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.23,"showMode":"Both"}}
Resizing
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"}}
Making horizontal
    Sidebar size = 120, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"},"horizontal":{"size":0.24,"showMode":"Both"}}
Resizing
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"},"horizontal":{"size":0.26,"showMode":"Both"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"},"horizontal":{"size":0.26,"showMode":"OnlyMain"}}
Recreating split view
    Sidebar size = 125, showMode = Both, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"Both"},"horizontal":{"size":0.26,"showMode":"OnlyMain"}}
Hiding sidebar
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"},"horizontal":{"size":0.26,"showMode":"OnlyMain"}}
Making horizontal
    Sidebar size = 0, showMode = OnlyMain, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"},"horizontal":{"size":0.26,"showMode":"OnlyMain"}}
Showing sidebar
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"},"horizontal":{"size":0.26,"showMode":"Both"}}
Recreating split view
    Sidebar size = 0, showMode = OnlyMain, vertical
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"},"horizontal":{"size":0.26,"showMode":"Both"}}
Making horizontal
    Sidebar size = 130, showMode = Both, horizontal
    Setting value: {"vertical":{"size":0.25,"showMode":"OnlyMain"},"horizontal":{"size":0.26,"showMode":"Both"}}
